- Done : create check list & pre-check list

+++ app/src/main/java/mobiqc/fpt/com/vn/mobiqc_v2kotlin_mvp/data/network/model/SubTeamModel.kt
@@ -4,4 +4,4 @@ package mobiqc.fpt.com.vn.mobiqc_v2kotlin_mvp.data.network.model
  * * Created by Anh Pham on 08/20/2018.     **
  * * Copyright (c) 2018 by FPT Telecom      **
  */
-data class SubTeamModel(val ResultDepID: String, val ResultCode: Float, val ResultSubID: String, val Ability: Float, val AbilityLeft: Float, val StatusBK: Float) : BaseModel()
+data class SubTeamModel(val ResultDepID: String, val ResultCode: Int, val ResultSubID: String, val Ability: Int, val AbilityLeft: Int, val StatusBK: Int) : BaseModel()
